requirements.txt
Flask==3.0.0
python-gnupg==0.5.2
python-dotenv==1.0.0
PyPDF2==3.0.1
reportlab==4.0.7
